HW13 Split server to grpc and http
