+++ hw12_13_14_15_calendar/internal/server/grpc/server.go
@@ -1,14 +1,11 @@
 package internalgrpc
 
 import (
-	"context"
 	"fmt"
 	"net"
-	"net/http"
 
 	grpc_middleware "github.com/grpc-ecosystem/go-grpc-middleware"
 	grpc_zap "github.com/grpc-ecosystem/go-grpc-middleware/logging/zap"
-	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
 	"github.com/leksss/hw-test/hw12_13_14_15_calendar/internal/domain/interfaces"
 	"github.com/leksss/hw-test/hw12_13_14_15_calendar/internal/infrastructure/config"
 	pb "github.com/leksss/hw-test/hw12_13_14_15_calendar/proto/protobuf"
@@ -18,7 +15,6 @@ import (
 
 type Server struct {
 	grpcAddr string
-	http     *http.Server
 	grpc     *grpc.Server
 	log      interfaces.Log
 	storage  interfaces.Storage
@@ -29,9 +25,6 @@ func NewServer(log interfaces.Log, config *config.Config, storage interfaces.Sto
 		log:      log,
 		storage:  storage,
 		grpcAddr: config.GRPCAddr.DSN(),
-		http: &http.Server{
-			Addr: config.HTTPAddr.DSN(),
-		},
 	}
 }
 
@@ -52,33 +45,6 @@ func (s *Server) StartGRPC() error {
 
 	s.log.Info(fmt.Sprintf("serving gRPC on %s", s.grpcAddr))
 	return s.grpc.Serve(lis)
-}
-
-func (s *Server) StartHTTPProxy() error {
-	conn, err := grpc.DialContext(
-		context.Background(),
-		s.grpcAddr,
-		grpc.WithBlock(),
-		grpc.WithInsecure(),
-	)
-	if err != nil {
-		s.log.Error("failed to dial server:", zap.Error(err))
-	}
-
-	gwMux := runtime.NewServeMux()
-	err = pb.RegisterEventServiceHandler(context.Background(), gwMux, conn)
-	if err != nil {
-		s.log.Error("failed to register gateway:", zap.Error(err))
-	}
-
-	s.http.Handler = loggingMiddleware(gwMux, s.log)
-	s.log.Info(fmt.Sprintf("serving gRPC-Gateway on %s", s.http.Addr))
-	return s.http.ListenAndServe()
-}
-
-func (s *Server) StopHTTPProxy(ctx context.Context) error {
-	s.log.Info("stopping HTTP proxy server...")
-	return s.http.Shutdown(ctx)
 }
 
 func (s *Server) StopGRPC() {
